Added hamburger menu

+++ src/components/HomePage.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import MemberForm from './MemberForm';
 import MemberFormAdmin from './MemberFormAdmin';
-import Hamburger from './Hamburger';
+import hamburgerSymbol from '../icons/hamburger.svg';
 
 import checkIfMemberExist from '../functions/checkIfMemberExist';
 import sendCodeByEmail from '../functions/sendCodeByEmail';
@@ -14,6 +14,9 @@ import '../styles/default.css';
 function HomePage() {
   const [loggedInUser, setLoggedInUser] = useState();
   const [loggedInUserRole, setLoggedInUserRole] = useState();
+
+  const [ display, setDisplay ] = useState( 'none' )
+   
 
 //*****************
 // Check if user exist in local storage.
@@ -35,35 +38,52 @@ function HomePage() {
   }, []);
 
   function submitLogOut(event) {
-    event.preventDefault();
+    // event.preventDefault();
     setLoggedInUser(null);
     setLoggedInUserRole(null);
     localStorage.clear();
   }
 
   async function submitLogin(event) {
-    event.preventDefault();
+    // event.preventDefault();
     const userEmailAddr = prompt('Skriv inn e-post adresse');
-    const [memberExist, , userRole] = await checkIfMemberExist('', userEmailAddr);
-    const randomCode = Math.floor(Math.random()*999999)+100001;
-    if (memberExist && (userRole === 'Admin' || userRole === 'Superuser')) {
-      sendCodeByEmail(userEmailAddr, randomCode);
+    if (userEmailAddr) {
+      const [memberExist, , userRole] = await checkIfMemberExist('', userEmailAddr);
+      const randomCode = Math.floor(Math.random()*999999)+100001;
+      if (memberExist && (userRole === 'Admin' || userRole === 'Superuser')) {
+        sendCodeByEmail(userEmailAddr, randomCode);
+      }
+      const userCode = prompt('Kode fra e-post: ');
+      if (parseInt(userCode) === randomCode) {
+        setLoggedInUser(userEmailAddr);
+        setLoggedInUserRole(userRole);
+        localStorage.setItem('user', JSON.stringify({username: userEmailAddr, userrole: userRole }));
+      }
+      else alert('Feil kode!');
     }
-    const userCode = prompt('Kode fra e-post: ');
-    if (parseInt(userCode) === randomCode) {
-      setLoggedInUser(userEmailAddr);
-      setLoggedInUserRole(userRole);
-      localStorage.setItem('user', JSON.stringify({username: userEmailAddr, userrole: userRole }));
-    }
-    else alert('Feil kode!');
   }
+
+  function hamburgerClick() {
+    if ( display === 'none' ) {
+        setDisplay( 'block' )
+    } else {
+        setDisplay( 'none' )
+    }
+}
   
   if (loggedInUser) {
     return (
       <div className='homepagetopdiv'>
-        <Hamburger loggedInuser={loggedInUser} />
-        <p>{loggedInUser}</p>
-        <button onClick={submitLogOut}>Logg ut</button>
+        <div className="homepageheaderdiv">
+          <img src={logo} alt="Logo" className="topimage" />
+          <div onClick={hamburgerClick} className="hamburgerdiv">
+              <img src={hamburgerSymbol} height='50' width='50' alt='menu-icon' />
+              <div className="hamburgerelementsdiv" style={{display:display}}>
+                  <p>{loggedInUser}</p>
+                  <button onClick={submitLogOut} >Logg ut</button>
+              </div>
+          </div>
+        </div>
         <MemberFormAdmin userLoggedIn={loggedInUser} userRole={loggedInUserRole} />
       </div>
     )
@@ -73,9 +93,13 @@ function HomePage() {
     <div className='homepagetopdiv'>
       <div className="homepageheaderdiv">
         <img src={logo} alt="Logo" className="topimage" />
-        <Hamburger loggedInUser='Ikke innlogget' />
+        <div onClick={hamburgerClick} className="hamburgerdiv">
+            <img src={hamburgerSymbol} height='50' width='50' alt='menu-icon' />
+            <div className="hamburgerelementsdiv" style={{display:display}}>
+                <button onClick={submitLogin} >Logg inn</button>
+            </div>
+        </div>
       </div>
-      <button onClick={submitLogin}>Logg inn</button>
       <MemberForm />
     </div>
   )
